[NFC][SSAF] Rename PointerFlowReachableAnalysis to UnsafeBufferReachableAnalysis
The previously named PointerFlowReachableAnalysis essentially
propagates unsafe buffers across a pointer flow graph. The pointer
flow analysis is a dependency rather than the subject
itself. Therefore, this commit renames and moves the analysis to
better reflect its purpose.
